// src/interfaces/member.ts
export interface Member {
    company?: string;
    department?: string;
    gender?: string;
    name: string;
    photo: string;
    picture?: string;
    points?: string;
    status?: string;
    last_online?: string;
    groups?:any;
    uid: string;
    user_goals?: string[];
    writing?: boolean;
}